Array Inversion Uploaded
Array Inversion using Recursion in O(n logn) time, where n is the size of the input.
Logic / Idea : It is piggy backed on mergesort algorithm. The function merge, where the merging of the two arrays take place - the left array and the right array, onto the a new output array , the copying of elements from left sub array is of no interest. But while copying the elements of right subarray, the number of remaining elements present in left subarray counts the number of split inversion.